+++ src/Make_dice.mak
@@ -27,6 +27,7 @@ LD = dcc
 
 SRC = \
 	arabic.c \
+	autocmd.c \
 	blowfish.c \
 	buffer.c \
 	charset.c \
@@ -84,6 +85,7 @@ SRC = \
 	version.c
 
 OBJ =	o/arabic.o \
+	o/autocmd.o \
 	o/blowfish.o \
 	o/buffer.o \
 	o/charset.o \
@@ -160,6 +162,8 @@ $(SYMS)  : vim.h globals.h keymap.h macros.h ascii.h term.h os_amiga.h structs.h
 ###########################################################################
 
 o/arabic.o:	arabic.c  $(SYMS)
+
+o/autocmd.o:	autocmd.c  $(SYMS)
 
 o/blowfish.o:	blowfish.c  $(SYMS)
